backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
requests==2.31.0
python-dotenv==1.0.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
pandas==2.1.4
numpy==1.25.2
schedule==1.2.0
apscheduler==3.10.4
pycryptodome==3.19.0
yfinance==0.2.28
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4